Refactor dashboard files
